ok filter render by 1click && search modularized in component && navbar always in / && ok professions showed in cards
