+++ client/src/App.js
@@ -9,24 +9,28 @@ import Login  from './components/Login';
 import Nosotros from './pages/Nosotros.jsx';
 import { ClientSpecificNeed } from './components/ClientSpecificNeed';
 import HomeProfessional from './components/HomeProfessional';
+import NavBar from './components/NavBar';
 
 function App() {
   return (
-    <Routes>
-      <Route exact path="/" element={<Home />} />
-      <Route exact path="/clients/:idClient" element={<ProfileClient />} />
-      <Route exact path="/professional/:id" element={<ProfileProfessional />} />
-      <Route exact path="/clients" element={<ProfileClient />} />
-      <Route exact path="/register" element={<Register />} />
-      <Route exact path="/nosotros" element={<Nosotros />} />
+    <>
+      <NavBar />
+      <Routes>
+        <Route exact path="/" element={<Home />} />
+        <Route exact path="/clients/:idClient" element={<ProfileClient />} />
+        <Route exact path="/professional/:id" element={<ProfileProfessional />} />
+        <Route exact path="/clients" element={<ProfileClient />} />
+        <Route exact path="/register" element={<Register />} />
+        <Route exact path="/nosotros" element={<Nosotros />} />
+          
+        <Route exact path="/homeprofessional" element={<HomeProfessional />} />
         
-      <Route exact path="/homeprofessional" element={<HomeProfessional />} />
-      
-      <Route exact path="/login" element={<Login />} />
-      <Route exact path="/necesidades" element={<ClientSpecificNeed />} />
+        <Route exact path="/login" element={<Login />} />
+        <Route exact path="/necesidades" element={<ClientSpecificNeed />} />
 
-      
-    </Routes>
+        
+      </Routes>
+    </>
   );
 }
