+++ client/src/components/styles/Search.modules.css
@@ -1,0 +1,23 @@
+
+.container__input{
+    display: flex;
+    align-items: center;
+    justify-content: space-around;
+    background-color: rgba(128, 128, 128, 0.068);
+    max-width: 60%;
+    height: 30px;
+    border-radius: 5px;
+    color: #707070;
+    padding-left: 5px;
+    /* margin-right: 5%; */
+} 
+
+.container__input--text{
+    border-radius: 10px;
+    border: none;
+    outline: none;
+    background-color: transparent;
+    font-family: 'Poppins', sans-serif;
+    color: #707070;
+    width: 80%;
+}
